# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TechniConnect - Electronics Technician & Shop Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shops:
    get:
      operationId: listShops
      tags: [shops]
      summary: List shops (with optional location filter)
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
            default: 50
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Shops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopListResponse"
    post:
      operationId: createShop
      tags: [shops]
      summary: Register a shop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
...
  /shops/{id}:
    get:
      operationId: getShop
      tags: [shops]
      summary: Get shop by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    put:
      operationId: updateShop
      tags: [shops]
      summary: Update shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    delete:
      operationId: deleteShop
      tags: [shops]
      summary: Delete shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /shops/{id}/verify:
    put:
      operationId: verifyShop
      tags: [shops]
      summary: Verify/unverify shop (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"

  # TECHNICIANS
  /technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List technicians (with optional location filter)
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
            default: 50
        - name: skill
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Technicians
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianListResponse"
    post:
      operationId: createTechnicianProfile
      tags: [technicians]
      summary: Register as technician
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
...
  /technicians/{id}:
    get:
      operationId: getTechnician
      tags: [technicians]
      summary: Get technician by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
    put:
      operationId: updateTechnician
      tags: [technicians]
      summary: Update technician profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
    delete:
      operationId: deleteTechnician
      tags: [technicians]
      summary: Delete technician profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /technicians/{id}/location:
    put:
      operationId: updateTechnicianLocation
      tags: [technicians]
      summary: Update technician live location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}/availability:
    put:
      operationId: updateTechnicianAvailability
      tags: [technicians]
      summary: Toggle technician availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}/verify:
    put:
      operationId: verifyTechnician
      tags: [technicians]
      summary: Verify/unverify technician (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"

  # PRODUCTS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LocationInput:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
      required: [lat, lng]
    AvailabilityInput:
      type: object
      properties:
        isAvailable:
          type: boolean
      required: [isAvailable]
...
    Shop:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        ratingCount:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        workingHours:
          type: ["string", "null"]
        createdAt:
          type: string
        ownerName:
          type: ["string", "null"]
      required: [id, ownerId, name, address, lat, lng, rating, ratingCount, isVerified, isActive, createdAt]
    ShopInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        phone:
          type: string
        email:
          type: string
        category:
          type: string
        images:
          type: array
          items:
            type: string
        workingHours:
          type: string
      required: [name, address, lat, lng]
    ShopUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        phone:
          type: string
        email:
          type: string
        category:
          type: string
        images:
          type: array
          items:
            type: string
        workingHours:
          type: string
        isActive:
          type: boolean
...
    ShopListResponse:
      type: object
      properties:
        shops:
          type: array
          items:
            $ref: "#/components/schemas/Shop"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [shops, total, page, limit]
    Technician:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        hourlyRate:
          type: number
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        isAvailable:
          type: boolean
        rating:
          type: number
        ratingCount:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        serviceRadius:
          type: integer
        profileImage:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, name, skills, experienceYears, hourlyRate, isAvailable, rating, ratingCount, isVerified, isActive, serviceRadius, createdAt]
    TechnicianInput:
      type: object
      properties:
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        hourlyRate:
          type: number
        serviceRadius:
          type: integer
        lat:
          type: number
        lng:
          type: number
      required: [skills, experienceYears, hourlyRate]
    TechnicianUpdate:
      type: object
      properties:
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        hourlyRate:
          type: number
        serviceRadius:
          type: integer
        isActive:
          type: boolean
...
    TechnicianListResponse:
      type: object
      properties:
        technicians:
          type: array
          items:
            $ref: "#/components/schemas/Technician"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [technicians, total, page, limit]
...