# lib/api-zod/src/generated/types/shopUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechniConnect - Electronics Technician & Shop Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ShopUpdate {
  name?: string;
  description?: string;
  address?: string;
  lat?: number;
  lng?: number;
  phone?: string;
  email?: string;
  category?: string;
  images?: string[];
  workingHours?: string;
  isActive?: boolean;
}